Add Context to DataAccess. Add Migrations. Change long id to int id. Add Mobile for User entity. Change BaseRepository to work with tables instead of static list of _entities. Change UserRepository Add Method. Add PackageRefrences to work with sql server. Add Connection string to appsettings and startup.

+++ DataAccess/Repositories/Interface/IRoleRepository.cs
@@ -2,7 +2,7 @@
 
 namespace DataAccess.Repositories.Interface
 {
-    public interface IRoleRepository : IBaseRepository<Role, long>
+    public interface IRoleRepository : IBaseRepository<Role, int>
     {
         void UpdateRule(Role oldRole, Role newRole);
     }
